+++ src/Date.cpp
@@ -64,8 +64,8 @@ string Date::toString(string format, string separator) {  // "DD.MM.YYYY"
     sort(index.begin(), index.end());
 
     ostringstream oss;
-    for (int i = 0; i < index.size(); i++)
-        oss << index[i].second << (i < index.size() - 1 ? separator : "");
+    for (int i = 0; i < (int)index.size(); i++)
+        oss << index[i].second << (i < (int)index.size() - 1 ? separator : "");
 
     return oss.str();
 }
